Collapse Renderer and Reader into one Type
